Refactor raw-handler back into index.js
